+++ packages/bluesky_cli/lib/bluesky_cli.dart
@@ -2,4 +2,4 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
-export 'package:bluesky_cli/src/command_runner.dart' show entryPoint;
+export 'package:bluesky_cli/src/command_runner.dart';