+++ packages/bluesky_cli/lib/src/command/commands.dart
@@ -2,6 +2,11 @@
 // Redistribution and use in source and binary forms, with or without
 // modification, are permitted provided the conditions.
 
+import 'actor/actor_suggestions.dart';
+import 'actor/actor_typeahead.dart';
+import 'actor/profile.dart';
+import 'actor/profiles.dart';
+import 'actor/search_actors.dart';
 import 'bsky_command.dart';
 import 'common/delete.dart';
 import 'feed/feed.dart';
@@ -19,6 +24,14 @@ import 'unspecced/popular.dart';
 
 final commonCommands = <BskyCommand>[
   DeleteCommand(),
+];
+
+final actorCommands = <BskyCommand>[
+  SearchActorsCommand(),
+  ProfileCommand(),
+  ProfilesCommand(),
+  ActorSuggestionsCommand(),
+  ActorTypeaheadCommand(),
 ];
 
 final feedCommands = <BskyCommand>[